backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.5.0
scikit-learn==1.2.2
nltk==3.8.1
numpy==1.24.3
pandas==1.5.3
